feat(give-page): initialize the give page mobile
