+++ src/assets/contentData/beliefs.tsx
