+++ src/assets/contentData/data.tsx
@@ -1,0 +1,19 @@
+export const data = {
+  ourStory: {
+    title: "Adonai Family",
+    storyTitle: "Who We Are",
+    storyDesc:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
+    getToKnow: "Get to know us",
+  },
+  theMission: {
+    title: "Who We Are",
+  },
+  give: {
+    title: "Generosity",
+    bodyTitle: "Thank you for giving",
+    bodyDesc:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+    givelify: "Give here",
+  },
+};
