--- src/assets/data.tsx
@@ -1,12 +1,0 @@
-export const data = {
-  ourStory: {
-    title: "Adonai Family",
-    storyTitle: "Who We Are",
-    storyDesc:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
-    getToKnow: "Get to know us",
-  },
-  theMission: {
-    title: "Who We Are",
-  },
-};
