+++ src/pages/_app.tsx
@@ -5,7 +5,7 @@ import Footer from "./Footer";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <div>
+    <div className="bg-regal-blue">
       <Navbar />
       <Component {...pageProps} />
       <Footer />
